// src/app/models/iuser-details-model.ts
export class IuserDetailsModel {

    
    FirstName: string;
    MiddleInitial: string;
    LastName: string;
    PhoneExt: string;
    //PhoneExt: number | null;
    AddressLine1: string;
    AddressLine2: string;
    State: string;
    StateCode: string;
    PostalCode: string;
    Country: number ;
    Fax: string;
    Cellno: string;
    OutlookEmail: string;
    QId: string | null;
    Answer: string;
    ClientSecurityQuestion: string;
    LoginId: string;
    login:string;
    Comment: string;
    ModifiedBy: string;
    IsAllowsedingSMS: boolean | null;
    SMTPEmail: string;
    SMTPPassword: string;
    SMTPAddress: string;
    SMTPPort: string;
    Email:string;
    RoleId:any;
    Phone:any;
    City:any;
    Password:any;
    UserId:number;
}